fix: use `castPtr` to avoid error on new text library
I had to deal with the `useAsPtr` in the new version of the text library, which seems to pass `Ptr Word8` instead of `Ptr UChar`.
It is probably safe to use `castPtr` for the same length, so I used it to solve the problem.
I think it's safe to assume that the version that depends on the old text will just be the same with cast.